docs: add some docs
