+++ sources/McQuery.Net/Data/Factories/RequestFactory.cs
@@ -1,11 +1,15 @@
 namespace McQuery.Net.Data.Factories;
 
+/// <summary>
+/// Implementation of <see cref="IRequestFactory"/>.
+/// </summary>
 internal class RequestFactory : IRequestFactory
 {
     private const byte HandshakeRequestTypeConst = 0x09;
     private const byte StatusRequestTypeConst = 0x00;
     private static readonly byte[] MagicConst = [0xfe, 0xfd];
 
+    /// <inheritdoc />
     public byte[] GetHandshakeRequest(SessionId sessionId)
     {
         using MemoryStream packetStream = new();
@@ -13,6 +17,7 @@ internal class RequestFactory : IRequestFactory
         return packetStream.ToArray();
     }
 
+    /// <inheritdoc />
     public byte[] GetBasicStatusRequest(Session session)
     {
         using MemoryStream packetStream = new();
@@ -20,6 +25,7 @@ internal class RequestFactory : IRequestFactory
         return packetStream.ToArray();
     }
 
+    /// <inheritdoc />
     public byte[] GetFullStatusRequest(Session session)
     {
         using MemoryStream packetStream = new();
@@ -35,7 +41,7 @@ internal class RequestFactory : IRequestFactory
         packetStream.Write(sessionId);
     }
 
-    private void FormBasicStatusRequest(Stream packetStream, Session session)
+    private static void FormBasicStatusRequest(Stream packetStream, Session session)
     {
         FormRequestHeader(packetStream, StatusRequestTypeConst, session.SessionId);
         packetStream.Write(session.Token);
