+++ sources/McQuery.Net/Data/Providers/ISessionIdProvider.cs
@@ -1,6 +1,13 @@
 namespace McQuery.Net.Data.Providers;
 
+/// <summary>
+/// Provides <see cref="SessionId"/> every time it's needed.
+/// </summary>
 internal interface ISessionIdProvider
 {
+    /// <summary>
+    /// Gets <see cref="SessionId"/>.
+    /// </summary>
+    /// <returns><see cref="SessionId"/>.</returns>
     SessionId Get();
 }
