+++ sources/McQuery.Net/Data/Providers/ISessionStorage.cs
@@ -2,7 +2,16 @@ using System.Net;
 
 namespace McQuery.Net.Data.Providers;
 
+/// <summary>
+/// Creates and stores <see cref="Session"/> objects.
+/// </summary>
 internal interface ISessionStorage: IDisposable
 {
-    Task<Session> GetAsync(IPEndPoint serverEndpoint, CancellationToken token = default);
+    /// <summary>
+    /// Gets stored session or acquire new.
+    /// </summary>
+    /// <param name="serverEndpoint"><see cref="IPEndPoint"/> to access Minecraft server by UDP.</param>
+    /// <param name="cancellationToken"><see cref="CancellationToken"/>.</param>
+    /// <returns><see cref="Session"/>.</returns>
+    Task<Session> GetAsync(IPEndPoint serverEndpoint, CancellationToken cancellationToken = default);
 }
